requirements.txt
arxiv==2.1.0
groq==0.9.0
openai==1.40.0
langchain==0.2.14
langchain-openai==0.1.22
langchain-community==0.2.12
faiss-cpu==1.9.0.post1
pyyaml==6.0.1
requests==2.32.3
pandas==2.2.2
matplotlib==3.9.0
seaborn==0.13.2
python-dotenv==1.0.1
rich==13.7.1
tiktoken==0.7.0
